+++ src/roq/hyperliquid/gateway/drop_copy.hpp
@@ -83,6 +83,7 @@ struct DropCopy final : public web::socket::Client::Handler, public protocol::js
   void operator()(Trace<protocol::json::OrderUpdates> const &) override;
   void operator()(Trace<protocol::json::Notification> const &) override;
   //
+  void operator()(Trace<protocol::json::ActionError> const &) override;
   void operator()(Trace<protocol::json::ActionOrder> const &) override;
   void operator()(Trace<protocol::json::ActionCancel> const &) override;