+++ src/roq/hyperliquid/gateway/market_data.hpp
@@ -86,6 +86,7 @@ struct MarketData final : public web::socket::Client::Handler, public protocol::
   void operator()(Trace<protocol::json::OrderUpdates> const &) override;
   void operator()(Trace<protocol::json::Notification> const &) override;
   //
+  void operator()(Trace<protocol::json::ActionError> const &) override;
   void operator()(Trace<protocol::json::ActionOrder> const &) override;
   void operator()(Trace<protocol::json::ActionCancel> const &) override;